consent-protocol/requirements.txt
# 🔐 Crypto + Signing
# SECURITY: Updated from 42.0.5 to fix CVE-2024-26130 (vulnerable OpenSSL in wheels)
cryptography>=44.0.0
# >=2.10.1 for mcp 1.26; [crypto] for JWT signing
pyjwt[crypto]>=2.10.1

# 🧪 Validation (use v2 binary wheel)
pydantic>=2.0,<3.0

# 🧬 Environment management
python-dotenv==1.0.1

# 🚀 Web Server
fastapi>=0.115.0
uvicorn>=0.34.0

# 🗃️ Database (Postgres async)
asyncpg>=0.29.0

# 🗄️ Supabase REST client (for migration)
# Pinned to avoid resolution-too-deep with default pip resolver
supabase==2.27.2

# ⚡ HTTP client for agents (e.g. Apple ID, APIs)
# >=0.27.1 for mcp 1.26; <0.29 for supabase 2.27
httpx>=0.27.1,<0.29

# 🔍 Code Quality & Linting moved to requirements-dev.txt

# 🚦 Rate Limiting & Real-time
slowapi>=0.1.9           # FastAPI rate limiting
sse-starlette>=2.1.0     # Server-Sent Events for consent notifications

# 📦 Optional OAuth / Apple ID agents
# SECURITY: Updated from 2.32.3 to fix CVE-2024-35195 (.netrc credentials leak)
requests>=2.32.4

# 🔌 MCP Server SDK (Model Context Protocol)
# Pinned to avoid resolution-too-deep with default pip resolver
mcp==1.26.0

# 🔥 Firebase Admin SDK (Identity)
# Pinned to avoid resolution-too-deep with default pip resolver
firebase-admin==7.1.0

# 🔧 Protocol Buffers & GRPC (Pinning to avoid resolution-too-deep errors)
# These versions are known to work well with google-api-core and firebase-admin
# 🔧 Protocol Buffers & GRPC (Stable Protobuf 6 Set)
# Anchoring to Protobuf 6.x to resolve conflicts between firebase-admin and google-adk
# grpcio 1.76+ provides pre-built wheels for Python 3.13 (see https://pypi.org/project/grpcio/)
# protobuf 6.x required by grpcio-status 1.76; compatible with firebase-admin and google-adk
protobuf>=6.31.1,<7.0.0
grpcio==1.76.0
grpcio-status==1.76.0
# 2.25.1+ required for protobuf 6.x; pinned to avoid resolution-too-deep
google-api-core==2.29.0
# proto-plus 1.24+ supports protobuf 6; pinned to avoid resolution-too-deep
proto-plus==1.27.0

# 🧠 Generative AI (ADK - Agent Development Kit)
# Pinned to avoid resolution-too-deep with default pip resolver
google-adk==1.23.0

# 🛠️ Utilities
wrapt>=1.14.0  # Required for Python 3.11+ compatibility
rsa>=4.9
google-auth>=2.14.1

# 📈 Market Data
yfinance>=0.2.36
